// fe/src/lib/api/generated/model/notificationResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export interface NotificationResponse {
  created_at: string;
  event_code: string;
  id: string;
  is_read: boolean;
  message: string;
  policy: string;
  read_at: string | null;
  target_id: string | null;
  target_type: string;
  title: string;
}